make discretization independent of row ordering
If we are unaware of the col values when we are computing spits we end
up with a 'virtual-split' that does not correspond to the choice of a
threshold because it is a split between two indices that have the same
col value. This fixes by skipping over potential cuts that occur at an
index where the col value does not change.